# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Xpress Ops Tower API",
    "version": "1.0.0",
    "description": "Production API for Xpress Ops Tower - Real-time operations command center",
    "contact": {
      "name": "Xpress Ops Team",
      "url": "https://xpress.ph"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.xpress.ph",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/alerts": {
      "get": {
        "summary": "(Frontend-called) GET /api/alerts",
        "description": "Route called by frontend - needs backend implementation",
        "tags": [
          "Alerts"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertsGetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get_api_alerts",
        "x-status": "baseline",
        "x-visibility": "internal",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/auth/mfa/challenge": {
      "get": {
        "summary": "(Frontend-called) GET /api/auth/mfa/challenge",
        "description": "Route called by frontend - needs backend implementation",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get_api_auth_mfa_challenge",
        "x-status": "baseline"
      }
    },
    "/api/auth/mfa/verify": {
      "get": {
        "summary": "(Frontend-called) GET /api/auth/mfa/verify",
        "description": "Route called by frontend - needs backend implementation",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get_api_auth_mfa_verify",
        "x-status": "baseline"
      }
    },
    "/api/bookings": {
      "get": {
        "summary": "(Frontend-called) GET /api/bookings",
        "description": "Route called by frontend - needs backend implementation",
        "tags": [
          "Bookings",
          "UAT"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingsGetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get_api_bookings",
        "x-status": "baseline",
        "x-visibility": "internal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-uat-approved": true,
        "x-promoted-at": "2025-09-07T07:19:08.092Z",
        "x-parked-at": "2025-09-07T07:37:46.228Z"
      }
    },
...
    "/api/pricing/events": {
      "get": {
        "summary": "(Frontend-called) GET /api/pricing/events",
        "description": "Route called by frontend - needs backend implementation",
        "tags": [
          "Pricing"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get_api_pricing_events",
        "x-status": "baseline"
      }
    },
    "/api/pricing/profiles": {
      "get": {
        "summary": "(Frontend-called) GET /api/pricing/profiles",
        "description": "Route called by frontend - needs backend implementation",
        "tags": [
          "Pricing"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingProfilesGetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get_api_pricing_profiles",
        "x-status": "baseline",
        "x-visibility": "internal",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/pricing/taxi-fares": {
      "get": {
        "summary": "(Frontend-called) GET /api/pricing/taxi-fares",
        "description": "Route called by frontend - needs backend implementation",
        "tags": [
          "Pricing"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingTaxiFaresGetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get_api_pricing_taxifares",
        "x-status": "baseline",
        "x-visibility": "internal",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/pricing/tnvs-fares": {
      "get": {
        "summary": "(Frontend-called) GET /api/pricing/tnvs-fares",
        "description": "Route called by frontend - needs backend implementation",
        "tags": [
          "Pricing"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingTnvsFaresGetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get_api_pricing_tnvsfares",
        "x-status": "baseline",
        "x-visibility": "internal",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/pricing/tolls": {
      "get": {
        "summary": "(Frontend-called) GET /api/pricing/tolls",
        "description": "Route called by frontend - needs backend implementation",
        "tags": [
          "Pricing"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get_api_pricing_tolls",
        "x-status": "baseline"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenericObject": {
        "type": "object",
        "additionalProperties": true
      },
...
      "BookingsGetResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
...
      "PricingProfilesGetResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
...
      "AlertsGetResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
      "PricingTaxiFaresGetResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
      "PricingTnvsFaresGetResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}